Move tagscheme tests to separate test file
